+++ examples/scenes/src/test_scenes.rs
@@ -2047,6 +2047,7 @@ mod impls {
             palette::css::BLACK,
             radius,
             std_dev,
+            false,
         );
     }